Assistant checkpoint: Corregir verificación de stock en facturas pendientes
Assistant generated file changes:
- API/Controllers/FacturacionController.cs: Corregir estructura de respuesta del endpoint VerificarStockFactura para que coincida exactamente con lo que espera el frontend
- GestionLlantera.Web/wwwroot/js/views/facturacion/facturacion.js: Corregir el procesamiento de la respuesta de verificación de stock para manejar mejor la estructura del API

---

User prompt:

No esta funcionando

+++ API/Controllers/FacturacionController.cs
@@ -953,7 +953,7 @@ namespace API.Controllers
                             productoId = detalle.ProductoId,
                             nombreProducto = detalle.NombreProducto,
                             descripcion = "Producto no encontrado en el sistema",
-                            precio = 0,
+                            precio = (decimal)0,
                             cantidadRequerida = detalle.Cantidad,
                             stockDisponible = 0,
                             problema = "Producto no encontrado",
